replaced old block with new SDK block
